Move WASM preparation tests out of implementation
Extract the WASM preparation helper tests into a sibling test module so
`wasm_preparation.rs` stays under the repository line limit.

Add test-only wrapper accessors and assert that override application wires
secrets and OAuth configuration when those runtime options are supplied.
Convert the shared WASM fixture to return `Result` so fixture failures
propagate through the tests instead of panicking.
